Trying to catch a bug

+++ app.coffee
@@ -1,6 +1,7 @@
 express = require "express"
 http    = require "http"
 sys     = require "sys"
+fs      = require "fs"
 process = require "child_process"
 app     = express()
 
@@ -8,9 +9,13 @@ app.configure ->
   app.set "port", 3333
 
 gitUpdate = (project, cb)->
-  command = "cd #{project}; git pull ; npm install; bower install --allow-root; cd -"
-  console.log "Executing #{command}"
-  process.exec command, cb
+  fs.exists project, (exists)->
+    if exists
+      command = "cd #{project}; git pull ; npm install; bower install --allow-root; cd -"
+      console.log "Executing #{command}"
+      process.exec command, cb
+    else
+      console.log "Incorrect Path #{exists}"
 
 
 # for development purpose only ( Hg update )
